+++ src/main.lua
@@ -1,5 +1,5 @@
 local function Main()
-    require 'unit_test.unit'
+    require 'unit_test.monster'
 end
 
 Main()